Add files via upload
These two files are related to the interactive coding tutorial by Joseph Lee Wei En. The article is titled 'Build your first Neural Network to predict house prices with Keras'. The URL is  https://medium.com/intuitive-deep-learning/build-your-first-neural-network-to-predict-house-prices-with-keras-eb5db60232c

The notebook file is not original work, rather a written duplication of the original author's program. The program intends to get the same results using the same code as the original author.